[FEAT] : deleted TradeTableViewCell
