+++ MLS/MLS/SharedViews/CommunityTableViewCell.swift
@@ -14,6 +14,8 @@ class CommunityTableViewCell: UITableViewCell {
 
     private let dateLabel = CustomLabel(text: "date", textColor: .gray, fontSize: 12)
     
+    private let tagLabel = CustomLabel(text: "tag", fontSize: 20)
+    
     private let titleLabel = CustomLabel(text: "title", fontSize: 20)
     
     private let upCountLabel = CustomLabel(text: "upCount", textColor: .gray, fontSize: 16)
@@ -22,7 +24,6 @@ class CommunityTableViewCell: UITableViewCell {
 
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-
         setUpConstraints()
     }
 
@@ -36,6 +37,7 @@ class CommunityTableViewCell: UITableViewCell {
     private func setUpConstraints() {
         addSubview(dateLabel)
         addSubview(titleLabel)
+        addSubview(tagLabel)
         addSubview(upCountLabel)
         
         dateLabel.snp.makeConstraints {
@@ -43,9 +45,15 @@ class CommunityTableViewCell: UITableViewCell {
             $0.leading.equalToSuperview().inset(Constants.defaults.horizontal)
         }
         
-        titleLabel.snp.makeConstraints {
+        tagLabel.snp.makeConstraints {
             $0.top.equalTo(dateLabel.snp.bottom).inset(-Constants.defaults.vertical)
             $0.leading.equalToSuperview().inset(Constants.defaults.horizontal)
+            $0.bottom.equalToSuperview().inset(Constants.defaults.vertical)
+        }
+        
+        titleLabel.snp.makeConstraints {
+            $0.top.equalTo(dateLabel.snp.bottom).inset(-Constants.defaults.vertical)
+            $0.leading.equalTo(tagLabel.snp.trailing).inset(-Constants.defaults.horizontal)
             $0.bottom.equalToSuperview().inset(Constants.defaults.vertical)
             $0.trailing.equalTo(upCountLabel.snp.leading).inset(Constants.defaults.horizontal)
         }
@@ -57,9 +65,24 @@ class CommunityTableViewCell: UITableViewCell {
         }
     }
     
-    func bind(date: String, title: String, upCount: String) {
-        dateLabel.text = date
+    func bind(tag: Constants.PostType, title: String, date: String, upCount: String) {
+        setUpNormal(tag: tag)
+        tagLabel.text = tag.rawValue
         titleLabel.text = title
+        dateLabel.text = date
         upCountLabel.text = upCount
+    }
+    
+    private func setUpNormal(tag: Constants.PostType) {
+        if tag == .normal {
+            tagLabel.isHidden = true
+            tagLabel.snp.makeConstraints {
+                $0.size.equalTo(0)
+            }
+            
+            titleLabel.snp.remakeConstraints {
+                $0.leading.equalToSuperview().inset(Constants.defaults.horizontal)
+            }
+        }
     }
 }
